[FIX] UI쪽 버그 및 score 타입변환 오류 수정(#27)

+++ LevelUpiOS/LevelUpiOS/Network/DTO/Submission/SubmissionResponse.swift
@@ -15,5 +15,5 @@ struct SubmissionResponse: Decodable {
 struct Submission: Decodable {
     let id: Int
     let examId: Int
-    let score: Int
+    let score: Double
 }
